codebase/functions/RIV/requirements.txt
paramiko==3.4.0
Pillow==10.3.0
torch==2.2.2
torchvision==0.17.2
